+++ Example/MultiSliderDemo/ViewController.swift
@@ -41,6 +41,7 @@ class ViewController: UIViewController {
         horizontalMultiSlider.tintColor = .purple
         horizontalMultiSlider.trackWidth = 32
         horizontalMultiSlider.showsThumbImageShadow = false
+        horizontalMultiSlider.valueLabelAlternatePosition = true
         horizontalMultiSlider.addTarget(self, action: #selector(sliderChanged), for: .valueChanged)
         view.addConstrainedSubview(horizontalMultiSlider, constrain: .leftMargin, .rightMargin, .bottomMargin)
         view.layoutMargins = UIEdgeInsets(top: 32, left: 32, bottom: 32, right: 32)