+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- add `valueLabelAlternatePosition` to allow more space for the value labels by alternating their positions for consecutive thumbs (eg top/bottom or leading/trailing).
+
 ## [2.0.0] - 2023-03-12
 
 ### Added